feature: rota de pontuacao
