+++ src/api/main.py
@@ -5,7 +5,7 @@ from src.views.ChaveamentoView import chaveamento_router
 from src.views.JuradoView import jurado_router
 from src.views.GraduacaoView import graduacao_router
 from fastapi.middleware.cors import CORSMiddleware
-from src.views.JogoView import jogo_router
+from src.views.PontuacaoView import pontuacao_router
 from uvicorn import run
 
 app = FastAPI()
@@ -36,4 +36,4 @@ app.include_router(competidor_router)
 app.include_router(chaveamento_router)
 app.include_router(jurado_router)
 app.include_router(graduacao_router)
-app.include_router(jogo_router)
+app.include_router(pontuacao_router)
