+++ src/schemas/PontuacaoSchema.py
@@ -1,0 +1,14 @@
+# from pydantic import BaseModel
+# import uuid
+#
+#
+# class PontuacaoBase(BaseModel):
+#     pontuacao_competidor_1: float
+#     pontuacao_competidor_2: float
+#     pontuacao_jogo: float
+#     id_jurado: str
+#     id_jogo: str
+#
+#
+# class PontuacaoRequest(PontuacaoBase):
+#     ...
